refactor(commands): extract /goal command into goal.rs
